backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
openai>=1.50.0
pandas>=2.2.0
python-multipart>=0.0.9
pypdf>=5.0.0
reportlab>=4.2.0
python-dotenv>=1.0.1
